Boetelijst/requirements.txt
Django==3.2
djangorestframework==3.14.0
django-cors-headers==4.0.0
drf_yasg==1.21.5
PyYAML==6.0